Remove strict validation of localized keys in desktop file reader
Making the validation more strict broke pretty much all AppImage builds on the KDE binary factory. Also, the validation is a nonfunctional aspect in this case, even if the keys' syntax was broken, the code could continue. I don't recall why I opted to implement such validation there in the first place. The strict validation should most likely be an optional feature (e.g., provided in a method validate()), but not done in the constructor.